+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/combat/fletching/FletchingFeature.java
@@ -1,0 +1,10 @@
+package insane96mcp.insanesurvivaloverhaul.module.combat.fletching;
+
+import insane96mcp.insanelib.core.feature.Feature;
+import insane96mcp.insanelib.core.feature.LoadFeature;
+import insane96mcp.insanesurvivaloverhaul.module.ISOModules;
+
+@LoadFeature(module = ISOModules.COMBAT, description = "Give an use to the fletching table, adding more arrows and better crafting")
+public class FletchingFeature extends Feature {
+    
+}